chore: refactor graph formula adapters to share common code [PT-186437311]
